Fixed FPS limit being enabled by default
There were some inconsistencies between the values set up in
`Client::initConfiguration` and those in `getConfigDefaults`. These
duplicates have now been removed.

For some of these settings the code getting the values had to be
adjusted to use getBoolValue, to actually rely on the provided default
instead of one provided as a parameter.

+++ src/gui/widgets/chattab.cpp
@@ -52,7 +52,7 @@ ChatTab::ChatTab(const std::string &name)
 
     mTextOutput = new BrowserBox(BrowserBox::AUTO_WRAP);
     mTextOutput->setWrapIndent(15);
-    mTextOutput->setMaxRows((int) config.getIntValue("ChatLogLength"));
+    mTextOutput->setMaxRows(config.getIntValue("ChatLogLength"));
     mTextOutput->setLinkHandler(chatWindow->mItemLinkHandler);
 
     mScrollArea = new ScrollArea(mTextOutput);
